+++ src/app/create/components/contentrequired.tsx
@@ -1,10 +1,9 @@
-/* eslint-disable no-param-reassign */
 /* eslint-disable react/jsx-props-no-spreading */
 import React, { useCallback, useMemo } from 'react';
 import { Controller, useFormContext, useWatch } from 'react-hook-form';
 
 import { TriangleAlert as TriangleAlertIcon } from 'lucide-react';
-import { BaseEditor, createEditor, Descendant, Editor } from 'slate';
+import { BaseEditor, createEditor, Descendant } from 'slate';
 import { HistoryEditor, withHistory } from 'slate-history';
 import {
   Editable,
@@ -22,7 +21,8 @@ import ImageTooltip from '@/components/TextEditor/ImageTooltip';
 import Leaf from '@/components/TextEditor/Leaf';
 import MarkButton from '@/components/TextEditor/MarkButton';
 import useEditorShortcuts from '@/components/TextEditor/hooks/useEditorShortcuts';
-import useImage, { type ImageElement } from '@/components/TextEditor/hooks/useImage';
+import useImage from '@/components/TextEditor/hooks/useImage';
+import useWithImages from '@/components/TextEditor/hooks/useWithImages';
 import { TEXT_EDITOR_BLOCK_ICON, TEXT_EDITOR_MARK_ICON } from '@/static/textEditorIcon';
 import countContentText from '@/utils/countContentText';
 
@@ -39,13 +39,6 @@ declare module 'slate' {
     Element: CustomElement;
   }
 }
-
-type WithImagesProps = {
-  editor: Editor;
-  isImageUrl: (url: string) => boolean;
-  insertImage: (editor: Editor, url: string) => void;
-};
-
 const MAX_CONTENT = 500;
 
 const initialValue: Descendant[] = [
@@ -55,41 +48,6 @@ const initialValue: Descendant[] = [
   },
 ];
 
-const withImages = ({ editor, isImageUrl, insertImage }: WithImagesProps) => {
-  const { isVoid, insertData } = editor;
-
-  editor.isVoid = (element: ImageElement) => {
-    return element.type === 'image' ? true : isVoid(element);
-  };
-
-  editor.insertData = (data: Editor) => {
-    const text = data.getData('text/plain');
-    const { files } = data;
-
-    if (files && files.length > 0) {
-      for (let i = 0; i < files.length; i += 1) {
-        const reader = new FileReader();
-        const [mime] = files[i].type.split('/');
-
-        if (mime === 'image') {
-          reader.addEventListener('load', () => {
-            const url = reader.result;
-            insertImage(editor, url as string);
-          });
-
-          reader.readAsDataURL(files[i]);
-        }
-      }
-    } else if (isImageUrl(text)) {
-      insertImage(editor, text);
-    } else {
-      insertData(data);
-    }
-  };
-
-  return editor;
-};
-
 function ContentRequired() {
   const {
     control,
@@ -98,6 +56,7 @@ function ContentRequired() {
   } = useFormContext<AuctionInputs>();
 
   const { isImageUrl, insertImage } = useImage();
+  const { withImages } = useWithImages();
 
   const contentRequiredValue = useWatch({ name: 'contentRequired', control });
   const contentRequired = contentRequiredValue || '0';
